fix(value): handle transfer-full, in arguments properly
This is done by making V8ToGIArgument() accept the transfer status and
react accordingly. I does not make the new argument  optional; this
forces me into thinking about the appropriate value to put in for
various callsites of this function.

The make the test added in the previous commit works.