fix(value): don't assume that zero-terminated array is of type gchar**
The array's element type can be arbitrary, so we need to check the value
with the correct type. The long switch and a template function is used
to do this job.

We can't just memcmp() with item_size, because one of the type is float,
and we can't assume a 0.0 float is the same as 0 integer.

A test, based on how I discovered this, is added.